backend/requirements.txt
# Core dependencies with specific versions to avoid conflicts
fastapi==0.104.1
uvicorn[standard]==0.24.0
httptools==0.6.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# WebSocket
python-socketio==5.10.0
python-engineio==4.8.0
websockets==12.0

# AI/ML
openai==1.3.7
langchain==0.0.340
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.3

# Other
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Additional dependencies
email-validator==2.1.0
PyJWT==2.8.0
langchain-anthropic==0.0.1
prometheus_client==0.19.0
structlog==23.2.0
qdrant-client==1.7.0
sentence-transformers==2.2.2